apps/api/pyproject.toml
[project]
name = "le-poulet-api"
version = "0.1.0"
description = "Le Poulet - City-wide chicken hunt game API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.111.0",
    "sqlmodel>=0.0.19",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "aioredis>=2.0.1",
    "httpx>=0.27.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sentry-sdk[fastapi]>=2.6.0",
    "python-multipart>=0.0.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "UP"]
target-version = "py312"

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
